Update installation instructions
When trying to install ecs-cli I had problems with the docker lib. It either mean the copy I had in my gopath was out of date, or docker had updated their libs more recently that we'd updated ecs-cli. It could well be both.

Document up a more repeatable installation process.